# packages/contracts/openapi/academic.json
{
  "components": {
    "schemas": {
...
      "ApiResponseAttendanceResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/AttendanceResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ApiResponseBinding": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/Binding"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ApiResponseBranchAttendanceStat": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/BranchAttendanceStat"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseClassGroupAttendanceStat": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ClassGroupAttendanceStat"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseLeaveResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/LeaveResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseListAttendanceResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/AttendanceResponse"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseListSchedule": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/Schedule"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseMapStringObject": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "additionalProperties": {},
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseScheduleResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ScheduleResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ApiResponseStudentAttendanceStat": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/StudentAttendanceStat"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseTeacherScheduleResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/TeacherScheduleResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "AttendanceResponse": {
        "properties": {
          "checkInAt": {
            "format": "date-time",
            "type": "string"
          },
          "checkInMethod": {
            "type": "string"
          },
          "checkOutAt": {
            "format": "date-time",
            "type": "string"
          },
          "classGroupName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lessonEndAt": {
            "format": "date-time",
            "type": "string"
          },
          "lessonId": {
            "type": "string"
          },
          "lessonStartAt": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "statusLabel": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          },
          "studentName": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "Binding": {
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "BranchAttendanceStat": {
        "properties": {
          "absent": {
            "format": "int32",
            "type": "integer"
          },
          "branchId": {
            "type": "string"
          },
          "classGroups": {
            "items": {
              "$ref": "#/components/schemas/ClassGroupAttendanceStat"
            },
            "type": "array"
          },
          "leave": {
            "format": "int32",
            "type": "integer"
          },
          "present": {
            "format": "int32",
            "type": "integer"
          },
          "rate": {
            "type": "number"
          },
          "total": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "CheckInRequest": {
        "properties": {
          "guardianId": {
            "type": "string"
          },
          "lessonId": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "qrCode": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "required": [
          "method"
        ],
        "type": "object"
      },
...
      "ClassGroupAttendanceStat": {
        "properties": {
          "absent": {
            "format": "int32",
            "type": "integer"
          },
          "classGroupId": {
            "type": "string"
          },
          "classGroupName": {
            "type": "string"
          },
          "leave": {
            "format": "int32",
            "type": "integer"
          },
          "present": {
            "format": "int32",
            "type": "integer"
          },
          "rate": {
            "type": "number"
          },
          "total": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "DaySchedule": {
        "properties": {
          "date": {
            "format": "date",
            "type": "string"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/ScheduleLessonItem"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "LeaveResponse": {
        "properties": {
          "approvedAt": {
            "format": "date-time",
            "type": "string"
          },
          "approvedBy": {
            "type": "string"
          },
          "branchId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "leaveEndDate": {
            "format": "date",
            "type": "string"
          },
          "leaveStartDate": {
            "format": "date",
            "type": "string"
          },
          "lessonId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "statusLabel": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          },
          "studentName": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "LessonCell": {
        "properties": {
          "capacity": {
            "format": "int32",
            "type": "integer"
          },
          "classRoomId": {
            "type": "string"
          },
          "classRoomName": {
            "type": "string"
          },
          "dayOfWeek": {
            "format": "int32",
            "type": "integer"
          },
          "endAt": {
            "format": "date-time",
            "type": "string"
          },
          "endMinute": {
            "format": "int32",
            "type": "integer"
          },
          "lessonId": {
            "type": "string"
          },
          "source": {
            "format": "int32",
            "type": "integer"
          },
          "startAt": {
            "format": "date-time",
            "type": "string"
          },
          "startMinute": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "studentCount": {
            "format": "int32",
            "type": "integer"
          },
          "teacherAvailabilityId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Schedule": {
        "properties": {
          "classGroupId": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "teacherId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ScheduleLessonItem": {
        "properties": {
          "classGroupId": {
            "type": "string"
          },
          "classGroupName": {
            "type": "string"
          },
          "classRoomId": {
            "type": "string"
          },
          "classRoomShortName": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          },
          "courseName": {
            "type": "string"
          },
          "endAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "startAt": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "teacherId": {
            "type": "string"
          },
          "teacherShortName": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ScheduleResponse": {
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/DaySchedule"
            },
            "type": "array"
          },
          "weekEnd": {
            "format": "date",
            "type": "string"
          },
          "weekStart": {
            "format": "date",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "StudentAttendanceStat": {
        "properties": {
          "absent": {
            "format": "int32",
            "type": "integer"
          },
          "leave": {
            "format": "int32",
            "type": "integer"
          },
          "present": {
            "format": "int32",
            "type": "integer"
          },
          "rate": {
            "type": "number"
          },
          "total": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "TeacherColumn": {
        "properties": {
          "branchId": {
            "type": "string"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonCell"
            },
            "type": "array"
          },
          "teacherId": {
            "type": "string"
          },
          "teacherName": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "TeacherScheduleResponse": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/TeacherColumn"
            },
            "type": "array"
          },
          "weekStart": {
            "format": "date",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/attendance/check-in": {
      "post": {
        "operationId": "checkIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAttendanceResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-controller"
        ]
      }
    },
...
    "/api/attendance/student/{studentId}": {
      "get": {
        "operationId": "studentHistory",
        "parameters": [
          {
            "in": "path",
            "name": "studentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListAttendanceResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-controller"
        ]
      }
    },
...
    "/api/leaves/{id}/approve": {
      "post": {
        "operationId": "approve_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLeaveResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "leave-controller"
        ]
      }
    },
...
    "/api/schedule/by-teacher": {
      "get": {
        "operationId": "byTeacher",
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "weekStart",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTeacherScheduleResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schedule-controller"
        ]
      }
    },
    "/api/schedule/my-week": {
      "get": {
        "operationId": "myWeek",
        "parameters": [
          {
            "in": "query",
            "name": "weekStart",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTeacherScheduleResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schedule-controller"
        ]
      }
    },
    "/api/schedule/week": {
      "get": {
        "operationId": "week",
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "weekStart",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseScheduleResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schedule-controller"
        ]
      }
    },
...
    "/api/stats/attendance/branch/{id}": {
      "get": {
        "operationId": "branch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBranchAttendanceStat"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-stat-controller"
        ]
      }
    },
    "/api/stats/attendance/class-group/{id}": {
      "get": {
        "operationId": "classGroup",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClassGroupAttendanceStat"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-stat-controller"
        ]
      }
    },
    "/api/stats/attendance/dashboard": {
      "get": {
        "operationId": "dashboard",
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMapStringObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-stat-controller"
        ]
      }
    },
    "/api/stats/attendance/student/{id}": {
      "get": {
        "operationId": "student",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentAttendanceStat"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-stat-controller"
        ]
      }
    },
...
    "/api/v1/academic/family/bindings/{id}/approve": {
      "post": {
        "operationId": "approve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBinding"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "family-controller"
        ]
      }
    },
...
    "/api/v1/academic/family/children/{id}/schedule": {
      "get": {
        "operationId": "schedule",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListSchedule"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "family-controller"
        ]
      }
    },
...
  }
}